+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Reduce the pause after card is drawn
 - Tweak tooltips
+- Remove the pause between cards played by Yogg-Sarron
 
 ### Fixed
 - Fix release URLs in Changelog